moved helper code into seperate package
