+++ rvn/libvirt.go
@@ -6,6 +6,7 @@ package rvn
 import (
 	"encoding/json"
 	"fmt"
+	librvnhelp "github.com/isi-lincoln/raven/rvnhelper"
 	"github.com/libvirt/libvirt-go"
 	xlibvirt "github.com/libvirt/libvirt-go-xml"
 	"io/ioutil"
@@ -550,7 +551,7 @@ func newDom(h *Host, t *Topo) *xlibvirt.Domain {
 			path := strings.Split(h.Image, "/")
 			baseImage += path[len(path)-1]
 		} else {
-			subPath, imageName, _ := ParseURL(parsedURL)
+			subPath, imageName, _ := librvnhelp.ParseURL(parsedURL)
 			baseImage += subPath + imageName
 		}
 		// this only leaves names, which default to deterlab and /rvn/img location
@@ -996,34 +997,4 @@ func cleanupRpcBind(net *libvirt.Network) {
 		return
 	}
 
-}
-
-// return a path, which we will create a directory tree with path[0]/path[1]/.../path[n]/image
-func ParseURL(parsedURL *url.URL) (path string, image string, err error) {
-	// Path is easier to use than RawPath
-	remoteFullPath := parsedURL.Path
-	splitPath := strings.Split(remoteFullPath, "/")
-	// get the image name, dont let user specify qcow2
-	// when rvn goes beyond qcow2, need to use correct format
-	image = splitPath[len(splitPath)-1]
-	// get the scheme used
-	// create necessary variables
-	var userName string
-	var hostName string
-	// now to create a directory tree from the path, omit scheme and opaque
-	if parsedURL.Opaque != "" {
-		return path, image, err
-	}
-	if parsedURL.User != nil {
-		userName = parsedURL.User.Username()
-		path = userName + "/"
-	}
-	if parsedURL.Host != "" {
-		hostName = parsedURL.Host
-		path = path + hostName + "/"
-	}
-	// ftp://user@host:/path will become user/host/path.../
-	pathMinusImage := strings.Join(splitPath[:len(splitPath)-1], "/")
-	path += pathMinusImage + "/"
-	return path, image, nil
 }
