+++ rvnhelper/imageParser.go
@@ -1,0 +1,116 @@
+package rvnhelper
+
+import (
+	"errors"
+	"io"
+	"log"
+	"net/http"
+	"net/url"
+	"os"
+	"os/exec"
+	"strings"
+)
+
+// https://gist.github.com/elazarl/5507969
+func CopyLocalFile(src, dst string) error {
+	s, err := os.Open(src)
+	if err != nil {
+		return err
+	}
+	defer s.Close()
+	d, err := os.Create(dst)
+	if err != nil {
+		return err
+	}
+	if _, err := io.Copy(d, s); err != nil {
+		d.Close()
+		return err
+	}
+	return d.Close()
+}
+
+func ValidateURL(input string) *url.URL {
+	parsedURL, _ := url.Parse(input)
+	return parsedURL
+}
+
+// return a path, which we will create a directory tree with path[0]/path[1]/.../path[n]/image
+func ParseURL(parsedURL *url.URL) (path string, image string, err error) {
+	// Path is easier to use than RawPath
+	remoteFullPath := parsedURL.Path
+	splitPath := strings.Split(remoteFullPath, "/")
+	// get the image name, dont let user specify qcow2
+	// when rvn goes beyond qcow2, need to use correct format
+	image = splitPath[len(splitPath)-1]
+	// get the scheme used
+	// create necessary variables
+	var userName string
+	var hostName string
+	// now to create a directory tree from the path, omit scheme and opaque
+	if parsedURL.Opaque != "" {
+		err = errors.New("Opaque URL not implemented")
+		return path, image, err
+	}
+	if parsedURL.User != nil {
+		userName = parsedURL.User.Username()
+		path = userName + "/"
+	}
+	if parsedURL.Host != "" {
+		hostName = parsedURL.Host
+		path = path + hostName + "/"
+	}
+	// ftp://user@host:/path will become user/host/path.../
+	pathMinusImage := strings.Join(splitPath[:len(splitPath)-1], "/")
+	path += pathMinusImage + "/"
+	return path, image, nil
+}
+
+func DownloadURL(parsedURL *url.URL, downloadPath string, imageName string) error {
+	URIScheme := parsedURL.Scheme
+	// if no scheme for downloading file is provided, default to https
+	// TODO: enforce HTTPS -- do not allow http, redirect
+	if URIScheme == "https" {
+		DownloadFile(downloadPath+imageName, parsedURL.String())
+	} else if URIScheme == "http" {
+		err := errors.New("http is not supported, please use https!")
+		return err
+	} else if URIScheme == "" {
+		DownloadFile(downloadPath+imageName, parsedURL.String())
+	} else {
+		err := errors.New(parsedURL.Scheme + " is not currently implemented!")
+		return err
+	}
+	return nil
+}
+
+// https://golangcode.com/download-a-file-from-a-url/
+func DownloadFile(filepath string, url string) error {
+	// Create the file
+	out, err := os.Create(filepath)
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+
+	// Get the data
+	resp, err := http.Get(url)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	// Write the body to file
+	_, err = io.Copy(out, resp.Body)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func CreateNetbootImage() error {
+	cmd := exec.Command("qemu-img", "create", "/var/rvn/img/netboot", "25G")
+	log.Printf("Creating netboot image")
+	err := cmd.Run()
+	return err
+}
